Add memory alignment size check in jmemmgr.c
SIMD architectures have different memory alignment requirements which
must be complied with to prevent performance - and more importantly -
security issues.

This patch introduces an explicit compile-time check to make sure
that ALIGN_SIZE is a multiple of 32 bytes for AVX2 and a multiple of
16 bytes for all other architectures.

Future Arm NEON performance patches will rely on this behaviour in
the memory manager for correctness.

Bug: 922430
Change-Id: I8cf0503fba301baa6505459e3308ba89e670cc26
Signed-off-by: Akash Srivastava <akashniki@gmail.com>