Add 32-byte memory alignment check for Arm NEON
ALIGN_SIZE is defined to be 32 bytes in order to accommodate AVX2's
alignment requirements. Arm NEON only requires 16-byte alignment but
software optimizations are possible if 32-byte alignment is assumed
for memory allocated by jmemmgr.c.

This patch introduces a compile-time check to make sure that
ALIGN_SIZE is a multiple of 32 bytes for Arm NEON systems.

Optimizations relying on this assumption for Arm NEON will follow
this CL.

Bug: 922430
Change-Id: I57f349eeffa704ec6c440f356a607964fc4fc8cc
Signed-off-by: Akash Srivastava <akashniki@gmail.com>